Use the single JS entrypoint
