+++ public/entrypoints.json
@@ -5,7 +5,7 @@
         "/bundles/terminal42tablelookupwizard/tablelookupwizard.b1302794.css"
       ],
       "js": [
-        "/bundles/terminal42tablelookupwizard/tablelookupwizard.129db3f8.js"
+        "/bundles/terminal42tablelookupwizard/tablelookupwizard.1713589d.js"
       ]
     }
   }
