+++ public/manifest.json
@@ -1,8 +1,8 @@
 {
   "tablelookupwizard.css": "/bundles/terminal42tablelookupwizard/tablelookupwizard.b1302794.css",
-  "tablelookupwizard.js": "/bundles/terminal42tablelookupwizard/tablelookupwizard.129db3f8.js",
+  "tablelookupwizard.js": "/bundles/terminal42tablelookupwizard/tablelookupwizard.1713589d.js",
   "images/info--dark.svg": "/bundles/terminal42tablelookupwizard/images/info--dark.d4af7fbc.svg",
   "images/info.svg": "/bundles/terminal42tablelookupwizard/images/info.23a08c7c.svg",
   "tablelookupwizard.b1302794.css.map": "/bundles/terminal42tablelookupwizard/tablelookupwizard.b1302794.css.map",
-  "tablelookupwizard.129db3f8.js.map": "/bundles/terminal42tablelookupwizard/tablelookupwizard.129db3f8.js.map"
+  "tablelookupwizard.1713589d.js.map": "/bundles/terminal42tablelookupwizard/tablelookupwizard.1713589d.js.map"
 }
